Summary_Analysis/requirements.txt
scipy
konlpy
python-docx
textrankr
pandas
olefile
pdfminer